# openapi.yaml
openapi: 3.0.0
info:
  title: 'Notification Dispatcher API'
  description: 'Multi-channel (SMS, Email, Push) notification system API. Requires API key via Authorization: Bearer <api_key>.'
  version: 1.0.0
servers:
  -
    url: /
    description: 'API base URL'
paths:
  /api/v1/notifications:
    get:
      tags:
        - Notifications
      summary: 'List notifications'
      description: 'List notifications with filters'
      operationId: 2eb7762229652cb9b0e4852a21a76707
      parameters:
        -
          name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - pending
              - queued
              - sent
              - failed
              - cancelled
        -
          name: channel
          in: query
          required: false
          schema:
            type: string
            enum:
              - sms
              - email
              - push
        -
          name: priority
          in: query
          required: false
          schema:
            type: string
            enum:
              - low
              - normal
              - high
        -
          name: batch_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'Paginated list of notifications'
    post:
      tags:
        - Notifications
      summary: 'Create notification(s)'
      description: 'Create single or batch notifications'
      operationId: b165da532b208d9abce42155fb42f10b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: 'Single notification object or { "notifications": [...] } for batch'
      responses:
        201:
          description: 'Notification(s) created'
        409:
          description: 'Duplicate idempotency key'
        422:
          description: 'Validation error'
  '/api/v1/notifications/{id}':
    get:
      tags:
        - Notifications
      summary: 'Get notification by ID'
      description: 'Get notification by ID'
      operationId: d62cf870f02e07f0cff2ead0d58d20ca
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'Notification details'
        404:
          description: 'Notification not found'
    delete:
      tags:
        - Notifications
      summary: 'Cancel notification'
      description: 'Cancel a notification'
      operationId: 2603f7c90eb081c5478938fa371abb83
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'Notification cancelled'
        404:
          description: 'Notification not found'
        422:
          description: 'Cannot cancel in current status'
  /api/v1/notifications/stats:
    get:
      tags:
        - Notifications
      summary: 'Get notification statistics'
      description: 'Get notification statistics'
      operationId: 7de3b6454ac39958eada58b71f7a8113
      parameters:
        -
          name: batch_id
          in: query
          required: false
          schema:
            description: 'Filter stats by batch (optional)'
            type: string
            format: uuid
      responses:
        200:
          description: 'Stats: total, by_status, by_channel, by_priority'
        422:
          description: 'Invalid batch_id if provided'
  /health:
    get:
      tags:
        - Observability
      summary: 'Health check'
      description: 'Comprehensive health check'
      operationId: bca2f77b836bb19d65e81e65c88823c2
      responses:
        200:
          description: 'All systems healthy'
        503:
          description: 'One or more systems unhealthy'
      security: []
  /api/v1/metrics:
    get:
      tags:
        - Observability
      summary: 'Get system metrics'
      description: 'Get comprehensive system metrics'
      operationId: 480b6c2e99aa892aa9c36beaa17c485a
      responses:
        200:
          description: 'Metrics: notifications, queue, database, cache, rate_limiting'
  /api/v1/templates:
    get:
      tags:
        - Templates
      summary: 'List templates'
      description: 'List all templates'
      operationId: 727fbbf4bfe60c743225fcb6806fcd8f
      parameters:
        -
          name: channel
          in: query
          required: false
          schema:
            type: string
            enum:
              - sms
              - email
              - push
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: 'Paginated list of templates'
    post:
      tags:
        - Templates
      summary: 'Create template'
      description: 'Create new template'
      operationId: c5bf09da2018218e034f7b479411b5fb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - slug
                - channel
                - content
      responses:
        201:
          description: 'Template created'
        409:
          description: 'Duplicate slug'
        422:
          description: 'Validation error'
  '/api/v1/templates/{id}':
    get:
      tags:
        - Templates
      summary: 'Get template by ID or slug'
      description: 'Get single template'
      operationId: 27822e06069ecb00984b3a9adf422d13
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Template details'
        404:
          description: 'Template not found'
    put:
      tags:
        - Templates
      summary: 'Update template'
      description: 'Update template'
      operationId: 5bb0a9a67505aa0c4259fe6ffa8c80aa
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema: []
      responses:
        200:
          description: 'Template updated'
        404:
          description: 'Template not found'
        422:
          description: 'Validation error'
    delete:
      tags:
        - Templates
      summary: 'Delete template'
      description: 'Delete template'
      operationId: c4070df786284ce285750994dcb4a548
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Template deleted'
        404:
          description: 'Template not found'
        409:
          description: 'Template in use'
  '/api/v1/templates/{id}/render':
    post:
      tags:
        - Templates
      summary: 'Render template with variables'
      description: 'Render template with variables'
      operationId: 80e732b114fe146089b772376e7439e6
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              properties:
                variables:
                  type: object
                  example:
                    name: John
              type: object
      responses:
        200:
          description: 'Rendered content'
        404:
          description: 'Template not found'
security:
  -
    bearerAuth: []
tags:
  -
    name: Notifications
    description: 'Notification CRUD and stats'
  -
    name: Templates
    description: 'Template CRUD and render'
  -
    name: Observability
    description: 'Health and metrics'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'API Key'
      description: 'API Key authentication. Use: Authorization: Bearer <api_key>'